Capital IQ Key Developments Headlines: Dataset Preparation
This contains of files that were used to prepare a dataset for further analysis. The outputs of these files (datasets that we will analyze) were too large to attach. Also, rerunning the code for the sentiment analysis for each headline will result in error. Must paste OpenAI Api key at specified location in order to run code fully.