Short-Term Dataset Exploration
The following files are explores the datasets for the Russell2000, S&P500, Nasdaq100, and Dow Jones 30 at the daily, weekly, and monthly levels. Through this analysis, we were able to understand the relationship that the amount of news articles has with stock returns and other traditional stock metrics. Key findings were those companies that have witnessed extreme returns (relative to the companies in the the respective index) draw more attention from the media. In addition to this, those companies that have news articles written about them also have higher volatility's of returns in the next period. We then create new datasets that will be used in regressions to attempt an predict directional movement of future returns.

- Russell2000 Daily Returns is Too Large to Attach. Must download data of daily returns of Russell2000 for 2023 to recreate