Add files via upload
Each rmd file performs three types of statistical models in attempt to predict the directional movement of a stocks returns in the future period. These three models are: logistic regression, lasso regression, and random forest. These predictions were constructed using machine learning applications to create predictions for each observation in each dataset, that will then be appended into a column for each of the respective datasets. The new product was used in the implementation of trading strategies.

-daily_russell_regression.csv is too large to attach. Must download Daily Russell Returns for 2023 and apply exploratory analysis to replicate analysis.
-standardized_daily_russell_regression.csv is too large to attach. Must download Daily Russell Returns for 2023 and apply exploratory analysis to replicate analysis.
-weekly_russell_regression.csv is too large to attach. Must download Daily Russell Returns for 2023 and apply exploratory analysis to replicate analysis.
-standardized_weekly_russell_regression.csv is too large to attach. Must download Daily Russell Returns for 2023 and apply exploratory analysis to replicate analysis.
-daily_sp500_regression.csv is too large to attach. Must perform Exploratory Analysis- S&P500 daily to replicate analysis.
-standard_daily_sp500_regression.csv is too large to attach. Must perform Exploratory Analysis- S&P500 daily to replicate analysis.