Short-Term Trading Strategies
The following files are trading strategies that were constructed to attempt to yield profits. We can see that several of these models outperform the market in the specified time span. It appears the random forest yield great returns when this column is implemented in the trading strategies. This leads us to be cautious when doing so as this indicates that this model is exposed to bias. Further fine tuning of the random forest would need to be performed before putting these strategies into practice with actual money. In addition to this, more prediction models should be performed to attempt to yield greater profits.

- Daily sp500 Trading Strategy is too large to attach. Must perform Daily SP500 Regression to get csv file to replicate this analysis
- Weekly Russell Trading Strategy is too large to attach. Must perform Weekly Russell Regression to get csv file to replicate this analysis
- Daily Russell Trading Strategy is too large to attach. Must perform Daily Russell Regression to get csv file to replicate this analysis