Long-Term Factor Creation
The purpose of the following files were to create four factors that will be used in our long-term trading strategy. The four factors are: culture emphasis, leadership retainment, employee income stability, and media popularity. Each of these four factors can be described in greater detail in the research paper. After the factors were joined together, we then proceeded with the long-term trading strategies.